fiscalDates, subscriptionDates
Refactor Periods to separate class. Create DateRange utility.
